refactor: Refactor coding app
- Create services.tasks for tasks services
- Update services.user_data
